Make gravity tolerance level less strict
The current tolerance levels for gravity are pretty strict. On my CV1, tilt
correction rarely kicks in because of this, even when the device stands
perfectly still. This change aims to (and in my experience, does) fix that by
being less strict on accelerometer checking while still checking the computed
average strictly.

The “* 2.0f” multiplier is largely arbitrary, and 1.5f seems enough in my case,
but much higher values did not have obvious ill effects either when I played
with this constant.