[Rift] Possibly more accurate Rift IMU handling
Use device-provided timestamps instead of fixed time steps. Also don't use
hard-coded values on number of samples per packet, which should make the
code more generic, provided the DK1 samples and timestamps have the same
meaning. I have experienced marginally less drift during a few tests, but
the difference was not wide enough for it to be really conclusive.

Overflow (should happen only once an hour on CV1, several times a second on
DK1) handling is also sub-optimal, as it justs defaults to a hardcoded sample
rate. This is not worse than the current code, though.

This has not been tested on DK1.