Add new device setting OHMD_IDS_KEEP_ON_AT_CLOSE.
This setting can be set to 1 to prevent OpenHMD from powering off
a device when it is closed with ohmd_close_device(). It defaults to
0 = try to power device off on close.

The flag so far applies to the HTC Vive. Powering off a device and
its display will cause the OS to detect the display as being unplugged.

This is, e.g., problematic in a X11 X-Server environment if one wants
to create a separate X-Screen for the video output of the HMD to gain
a performance boost and reduction in motion-to-photon latency, and to
avoid potential tearing artifacts. For best performance, the graphics
card should use page-flipping, but this only works if the HMD output
window covers the full X-Screen, so this can only be achieved if one
creates a separate X-Screen for the HMD. Separate X-Screens only work
if the HMD display is active (= "connected") while the X-Server starts,
otherwise such a separate X-Screen is discarded, iow. the separate
X-Screen is not hot-pluggable.

This flag allows to run a little helper application which opens the
HMD, e.g. during system boot, then closes it again, just to turn on
the display. Then one can (re-)start the X-Server, getting the extra
X-Screen for the HMD, and then use a VR application enjoying the
better performance and timing of this setup.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>