vrtek: update WVR2 support to assume user has rotated display
Today, OpenHMD assumes that the user has rotated their display if the
HMD they're using has it's panel rotated. In the case of the WVR2 that
means we should be advertising 2560x1440 because that's what the
display will act as once the user does:
  xrandr --output HDMI-A-1 --rotate left
  (or whatever the case may be for their output)

While we're at it, let's also let each display type set their own
distortion and aberration coefficients. Once they're tuned they'll
probably be different after all.