vrtek: add support for using our own sensor fusion
This patch adds support for using OpenHMD's own sensor fusion
implementation to generate the orientation quaternion. It uses the
gyroscope and accelerometer readings found in the HMD data packet to
do so. The gyro offset values used here were found experimentally. They
are the mean of a large number of values reported with the headset
stationary (~20,000 samples). In the future it would be nice if we
could query them from the headset itself, especially as they're
somewhat temperature dependent.

For completeness we keep the ability to use the orientation quaternion
provided by the HMD. This can be used by setting the
OHMD_VRTEK_USE_DIRECT_QUATERNION environment variable to "true". The
default behavior is to use OpenHMD's sensor fusion, however.

(While we're at it let's also be a little more consistent with
formatting - when we have an else put it on the same line as the
closing brace of the above if statement.)