Add initial Rift S driver
This is a preliminary 3DOF driver for the Rift S headset and
controllers. Rotational IMU tracking and controller
controls are supported. It would also be possible to
publish more from the controllers with additional OpenHMD API,
such as capsense touch information for each button/trigger.

Distortion parameters are taken from the CV1 and need improving,
but aren't completely wrong.

Partially fixes #225