Initial 3DOF Oculus Rift S driver (#299)
* Add initial Rift S driver

This is a preliminary 3DOF driver for the Rift S headset and
controllers. Rotational IMU tracking and controller
controls are supported. It would also be possible to
publish more from the controllers with additional OpenHMD API,
such as capsense touch information for each button/trigger.

Distortion parameters are taken from the CV1 and need improving,
but aren't completely wrong.

Partially fixes #225

* hid.h: Add include guards to the header

* rift-s: Fix CMake build

* rift-s: Use MSVC-compatible packed structs

Use #pragma pack() instead of the GCC packed
attributes to fix MSVC compilation

* rift-s: Fix compile warnings with MSVC

Fix implicit conversion and format string warnings with MSVC

* rift-s: Fix controller report handling on Windows

Windows seems to supply a 64-byte HID report for the controller
report, so handle that. Also handle invalid reports better in
general by not accessing uninitialised memory or asserting

* rift-s: Fix trigger, grip and button ids

Re-scale the trigger and grip reports to be 0.0 for released to 1.0
for full, and switch the left/right controllers so the buttons match.