Use PROJECT_SOURCE_DIR for CMAKE_MODULE_PATH
The error that occurs using CMAKE_SOURCE_DIR for CMAKE_MODULE_PATH is, that the HIDAPI  package could not be found due to the missing config file FindHIDAPI.cmake. So it is not possible to use OpenHMD in other CMake projects. Using  PROJECT_SOURCE_DIR instead of CMAKE_SOURCE_DIR solves the issue.

For more details see the CMake documentation:
CMAKE_SOURCE_DIR is the path to the top level of the source tree (https://cmake.org/cmake/help/v3.1/variable/CMAKE_SOURCE_DIR.html). It differs, if OpenHMD is included to other projects.
PROJECT_SOURCE_DIR is the top level source directory for the current project. This is the source directory of the most recent project() command. (https://cmake.org/cmake/help/v3.1/variable/PROJECT_SOURCE_DIR.html). Due to the project is defined in the top level CMakeLists.txt the correct path will be used to find the modules.