[wmr] Fix type of wmr_config_header field
The calibration header contains a 16 bit version field, which we are
incorrectly parsing as part of the json_start field. Presumably this has
gone unnoticed because the header version field is always set to 0 in
practice.

Change type of json_start field from uint32_t to uint16_t, and insert a
new field uint16_t header_version.