Updated Simple to allow polling of Virtual Devices.
Some VR Devices' data is routed through another device.  This commit
updates the simple program to allow polling of such devices.

This commit also adds a feature for allowing the user to select the
device that is polled via the command line first showing the device
ids and then promting the user to select the device to be polled.