Handled NOLO STATION connected.
When the nolo base station is connected, if it is the first
device returned by the hid_enumerate, than OpenHMD exits
because the STATION is the wrong device.

This code correctly iterates through the linked list ignoring
the STATION if it is connected and ensures that the other NOLO
devices are registered.