Normalized Trackpad and fixed controller assign.
The trackpad reports values with the range of [0,255] for x and y axis.
Monado expects a range of [-1, 1] for both axis.  This commit normalizes the
output for the trackpad to [-1, 1].

The controller y and z axis were assigned to the incorrect values (swapped).
This commit only changes the y and z axis for the controllers as the
head tracker appears to be correct.