Fixed the gyo and acceleration variable swap.
The gyo and acceleration data from nolo was reversed.  This caused
the code to swap the variable names.  It could be that in previous
nolo firmware the data formatting was different than the latest firmware?

This commit fixes the gyo and acceleration when retreiving those
values from the nolo controllers.