added rudimentary support for the Pimax 4K HMD
This commit merges with minimal cleanup and adjustments the branch "dev-pimax" by TheOnlyJoey.
Raw USB/HID captures of a Pimax 4k were analyzed. The significant changes were that only
HID responses of type "11" are relevant and can be triggered with HID request "17", and
they run into a timeout after ~3 sec if no other "17" request is sent. The headset can still
crash/stop sending data. Sometimes it resumes after a while. Otherwise, it needs to be replugged.

During testing, gyro drifting was noticed. So far, this was fixed with magic numbers gathered
with external python scripts "grafdigital/OpenHmdPimaxTools" that may only work with the headset
the data was pulled from. Should we look into setup functions or make these values
configurable through files?

Lens correction needs improvement. This is the first draft to get passable graphical output.
Head tracking is passable but lags a bit, but this could be due to the device.

Functionality was validated with "openglexample".

Still missing is the utilisation of magnetometer data.