'File' HMD driver for integration with homebrew setups
This is sort of similar in spirit to Monado having an 'arduino' driver (presumably for SlimeVR).

However, this driver aims to allow a user to very quickly homebrew a 3DOF driver for a 'pure HMD' arrangement.

In particular, I am aiming to use this driver with Moonlight/Sunshine to stream to a Google Cardboard system.

Configuration is via environment variable to avoid the problem the "external" driver has (it's always present, unmarked, and needs to be ignored by end-user applications in practice).